fix: Resolve type mismatches and undefined function in ParameterTuningSearch
Fixed critical bugs that prevented parameter tuning from completing:

1. Type mismatch in update_best_attempt!:
   - Changed ppm_errs parameter from Vector{Float32} to Vector{Float64}
   - Updated IterationState.best_ppm_errs field to match
   - fit_mass_err_model returns Float64 vectors, now properly handled

2. Undefined function setRtConversionModel!:
   - Replaced with correct function setRtIrtMap!
   - Fixed argument order (model, then ms_file_idx)

These fixes allow the best attempt fallback mechanism to work correctly
when parameter tuning doesn't achieve full convergence.

+++ src/Routines/SearchDIA/SearchMethods/ParameterTuningSearch/types.jl
@@ -151,7 +151,7 @@ mutable struct IterationState
     best_psm_count::Int64
     best_mass_error_model::Union{Nothing, MassErrorModel}
     best_rt_model::Union{Nothing, Tuple{SplineRtConversionModel, Vector{Float32}, Vector{Float32}, Float32}}
-    best_ppm_errs::Union{Nothing, Vector{Float32}}
+    best_ppm_errs::Union{Nothing, Vector{Float64}}
     best_phase::Int64
     best_score::UInt8
     best_iteration::Int64
